refactor(dashboard): colocate single-consumer components with their routes
Move 4 single-consumer components out of the flat src/client/ui/components/
folder into route-local _components/ folders, following TanStack Router's
_-prefix convention (the router plugin ignores _-prefixed dirs, so they
aren't picked up as routes):

- edit-agent-dialog.tsx        -> routes/agents//_components/
- create-shared-env-set-dialog  -> routes/shared-env-sets/_components/
- edit-shared-env-set-dialog    -> routes/shared-env-sets/_components/
- env-var-dialog.tsx            -> routes/shared-env-sets/_components/

Each move updates a single import line in its consumer route file to use a
relative ./_components/ path. Moved files keep their @/-aliased imports
unchanged (e.g. edit-agent-dialog still imports code-editor from the
global components/ folder, since code-editor is multi-consumer).

The flat components/ folder drops from 9 to 5 components. app-sidebar and
copy-button stay as true global shared infra; agent-config-chat, code-editor,
and phase-badge stay global since they have multiple consumers.

Document the _components/ convention + promote-back rule in AGENTS.md: when a
second consumer appears, promote the component back to src/client/ui/components/.

Verified: typecheck clean, lint 0 errors (2 pre-existing
react-hooks/exhaustive-deps warnings unchanged by the move).
