+++ apps/dashboard/AGENTS.md
@@ -18,3 +18,4 @@ Run from the repo root via pnpm filter, or from this directory directly.
 - Use case classes compose mixins from `src/server/usecases/mixin.ts` (`BaseUseCase`, `HermeumConfigLoadable`, `OwnershipGuarded`) over injected `Runtime` (persistence) and `FileAdaptor` (docs) interfaces; concrete adaptors live in `src/server/infras/`. Inject mock adaptors in tests rather than relying on the defaults.
 - Field semantics for the chat agent live in tool input-schema `.describe()` texts (Zod), not in the system prompt — see the header comment on `AGENT_CONFIG_CHAT_SYSTEM_PROMPT` in `src/server/usecases/chat.ts`.
 - New chat tools follow the `readDocument` precedent: embed a lightweight list (names/ids + descriptions) in the system prompt up front via a `build*List()` method, and add a `read*` server-executed tool for fetching richer per-item detail on demand. Keep the prompt lean — batch lists into the prompt, batch detail calls into one tool invocation.
+- Route-scoped components live in a `_components/` folder next to the consuming route (e.g. `routes/agents/$id/_components/`). The TanStack Router plugin ignores `_`-prefixed folders, so they aren't picked up as routes. Promote a component back to `src/client/ui/components/` the moment a second consumer appears — otherwise route-local duplicates accumulate.
